refactor: GameRoom 타입 변경, serverRooms를 variables로 분리
