+++ backend/src/db.ts
@@ -1,10 +1,10 @@
 import mongoose from 'mongoose';
 
-const URI = process.env.MONGO_URI || '';
+const URI = process.env.MONGO_URI as string;
 
 mongoose.connect(URI, {});
 
 const db = mongoose.connection;
 
 db.once('open', () => console.log('mongoose connected'));
-db.on('error', (error) => console.log('DB Error', error));
+db.on('error', (error) => console.error('DB Error', error));
