+++ backend/src/types/GameRoom.d.ts
@@ -5,6 +5,8 @@ export type GameRoom = {
   players: { [socketId: string]: Player };
   title: string;
   playListId: string;
+  playlistName: string;
+  hashtags: string[];
   hasPassword: boolean;
   skip: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8;
   timePerProblem: 10 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90;
