+++ backend/src/variables/serverRooms.ts
@@ -1,0 +1,5 @@
+import { ServerRoom } from '../types/ServerRoom';
+
+const serverRooms: { [uuid: string]: ServerRoom } = {};
+
+export default serverRooms;
